+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The types of changes are:
 * Added taxonomy page to UI [#902](https://github.com/ethyca/fides/pull/902)
   * Added a nested accordion component for displaying taxonomy data [#910](https://github.com/ethyca/fides/pull/910)
 * Add lru cache to get_config [927](https://github.com/ethyca/fides/pull/927)
+* Add support for deprecated API config values [#959](https://github.com/ethyca/fides/pull/959)
 * `fides` is now an alias for `fidesctl` as a CLI entrypoint [#926](https://github.com/ethyca/fides/pull/926)
 * Add user auth routes [929](https://github.com/ethyca/fides/pull/929)
 * Bump fideslib to 3.0.1 and remove patch code[931](https://github.com/ethyca/fides/pull/931)